fix #35
When trying to use ArrayAdapter and dynamically loading new items was
leading to break the swipe to dismiss functionality because the internal
items counter was never refresh.

I've somehow reverse/fork the notifyDataSetChanged() if the BaseAdapter
is an ArrayAdapter.

At this point the google cards example is working with dynamic loading
of additional items using a custom version of OnScrollListener (items
are added when the scroll reach some threshold).

I've also tested the original example apk : so far so good.

I've put some Log.d() to help figure out the problem.